+++ src/components/Clipboard.react.js
@@ -3,6 +3,8 @@ import PropTypes from 'prop-types';
 import {FontAwesomeIcon} from '@fortawesome/react-fontawesome';
 import {faCopy, faCheckCircle} from '@fortawesome/free-regular-svg-icons';
 
+const clipboardAPI = navigator.clipboard;
+
 /**
  * The Clipboard component copies text to the clipboard
  */
@@ -13,13 +15,29 @@ export default class Clipboard extends React.Component {
         this.copyToClipboard = this.copyToClipboard.bind(this);
         this.copySuccess = this.copySuccess.bind(this);
         this.getText = this.getText.bind(this);
+        this.stringifyId = this.stringifyId.bind(this);
         this.state = {
             copied: false,
-            hasNavigator: true,
         };
     }
 
+    // stringifies object ids used in pattern matching callbacks
+    stringifyId(id) {
+        if (typeof id !== 'object') {
+            return id;
+        }
+        const stringifyVal = v => (v && v.wild) || JSON.stringify(v);
+        const parts = Object.keys(id)
+            .sort()
+            .map(k => JSON.stringify(k) + ':' + stringifyVal(id[k]));
+        return '{' + parts.join(',') + '}';
+    }
+
     copySuccess() {
+        this.props.setProps({
+            n_clicks: this.props.n_clicks + 1,
+        });
+
         const showCopiedIcon = 1000;
         this.setState({copied: true}, () => {
             setTimeout(() => {
@@ -30,30 +48,33 @@ export default class Clipboard extends React.Component {
 
     getText() {
         // get the inner text.  If none, use the content of the value param
-        var text = document.getElementById(this.props.target_id).innerText;
-        if (text === undefined || text === '') {
-            text = document.getElementById(this.props.target_id).value;
-            text = text === undefined ? '' : text;
+        const id = this.stringifyId(this.props.target_id);
+        const target = document.getElementById(id);
+        let text = target.innerText;
+        if (!text) {
+            text = target.value;
+            text = text === undefined ? null : text;
         }
         return text;
     }
 
     copyToClipboard() {
-        const text = this.getText();
-        navigator.clipboard.writeText(text).then(
-            () => {
-                this.copySuccess();
-            },
-            function() {
+        let text = this.props.text;
+        if (this.props.target_id) {
+            text = this.getText();
+        }
+        if (text) {
+            clipboardAPI.writeText(text).then(this.copySuccess, function() {
                 alert('copy error');
-            }
-        );
+            });
+        } else {
+            this.copySuccess();
+        }
     }
 
     componentDidMount() {
-        if (!navigator.clipboard) {
-            this.setState({hasNavigator: false});
-            alert('Copy to clipboard not available with this browser');
+        if (!clipboardAPI) {
+            console.warn('Copy to clipboard not available with this browser'); // eslint-disable-line no-console
         }
     }
 
@@ -63,7 +84,7 @@ export default class Clipboard extends React.Component {
         const copiedIcon = <FontAwesomeIcon icon={faCheckCircle} />;
         const btnIcon = this.state.copied ? copiedIcon : copyIcon;
 
-        return this.state.hasNavigator ? (
+        return clipboardAPI ? (
             <div id={id} title={title} style={style} className={className}>
                 <i onClick={() => this.copyToClipboard()}>{btnIcon}</i>
             </div>
@@ -71,7 +92,11 @@ export default class Clipboard extends React.Component {
     }
 }
 
-Clipboard.defaultProps = {};
+Clipboard.defaultProps = {
+    text: null,
+    target_id: null,
+    n_clicks: 0,
+};
 
 Clipboard.propTypes = {
     /**
@@ -80,11 +105,25 @@ Clipboard.propTypes = {
     id: PropTypes.string,
 
     /**
-     * id of target component containing text to copy to the clipboard.
-     *  The inner text of the `children` prop will be copied to the clipboard.  If none, then the text from the
+     * The id of target component containing text to copy to the clipboard.
+     * The inner text of the `children` prop will be copied to the clipboard.  If none, then the text from the
      *  `value` prop will be copied.
      */
-    target_id: PropTypes.string.isRequired,
+    target_id: PropTypes.oneOfType([PropTypes.string, PropTypes.object]),
+
+    /**
+     * The text to  be copied to the clipboard if the `target_id` is None.
+     * If the `target_id` is None and `text` is None, then the
+     * icon copy animation will display, but nothing is copied to the clipboard.  This
+     * allows for the use of other copy to clipboard methods in a callback, such as
+     *  `pandas.DataFrame.to_clipboard`.
+     */
+    text: PropTypes.string,
+
+    /**
+     * The number of times copy button was clicked
+     */
+    n_clicks: PropTypes.number,
 
     /**
      * The text shown as a tooltip when hovering over the copy icon.
@@ -102,8 +141,7 @@ Clipboard.propTypes = {
     className: PropTypes.string,
 
     /**
-     * Dash-assigned callback that should be called to report property changes
-     * to Dash, to make them available for callbacks.
+     * Dash-assigned callback that gets fired when the value changes.
      */
     setProps: PropTypes.func,
 };